refactor(Channel package): Moved channel utilities into a util package and removed the extra classes for email address handling.
