+++ src/main/java/com/synopsys/integration/alert/channel/util/ChannelEventManager.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.event;
+package com.synopsys.integration.alert.channel.util;
 
 import java.util.Map;
 import java.util.UUID;
