+++ src/main/java/com/synopsys/integration/alert/channel/util/ChannelFreemarkerTemplatingService.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel;
+package com.synopsys.integration.alert.channel.util;
 
 import java.io.File;
 import java.io.IOException;
