+++ src/main/java/com/synopsys/integration/alert/channel/util/ChannelRestConnectionFactory.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.rest;
+package com.synopsys.integration.alert.channel.util;
 
 import java.util.Optional;
