+++ src/main/java/com/synopsys/integration/alert/provider/polaris/PolarisProvider.java
@@ -34,7 +34,6 @@ import com.synopsys.integration.alert.common.provider.Provider;
 import com.synopsys.integration.alert.common.workflow.MessageContentCollector;
 import com.synopsys.integration.alert.common.workflow.task.ScheduledTask;
 import com.synopsys.integration.alert.common.workflow.task.TaskManager;
-import com.synopsys.integration.alert.provider.DefaultEmailHandler;
 import com.synopsys.integration.alert.provider.polaris.descriptor.PolarisContent;
 import com.synopsys.integration.alert.provider.polaris.tasks.PolarisProjectSyncTask;
 import com.synopsys.integration.polaris.common.rest.AccessTokenPolarisHttpClient;
@@ -51,9 +50,9 @@ public class PolarisProvider extends Provider {
     private final ObjectFactory<PolarisCollector> polarisCollector;
 
     @Autowired
-    public PolarisProvider(final TaskManager taskManager, final PolarisProjectSyncTask projectSyncTask, final PolarisProperties polarisProperties, final PolarisContent polarisContent, final DefaultEmailHandler defaultEmailHandler,
+    public PolarisProvider(final TaskManager taskManager, final PolarisProjectSyncTask projectSyncTask, final PolarisProperties polarisProperties, final PolarisContent polarisContent,
         final ObjectFactory<PolarisCollector> polarisCollector) {
-        super(PolarisProvider.COMPONENT_NAME, polarisContent, defaultEmailHandler);
+        super(PolarisProvider.COMPONENT_NAME, polarisContent);
         this.taskManager = taskManager;
         this.projectSyncTask = projectSyncTask;
         this.polarisProperties = polarisProperties;
