+++ src/test/java/com/synopsys/integration/alert/channel/email/EmailChannelChannelDescriptorTestIT.java
@@ -38,7 +38,6 @@ import com.synopsys.integration.alert.database.api.DefaultProviderDataAccessor;
 import com.synopsys.integration.alert.database.provider.user.ProviderUserEntity;
 import com.synopsys.integration.alert.database.provider.user.ProviderUserRepository;
 import com.synopsys.integration.alert.mock.MockConfigurationModelFactory;
-import com.synopsys.integration.alert.provider.DefaultEmailHandler;
 import com.synopsys.integration.alert.provider.blackduck.BlackDuckProvider;
 import com.synopsys.integration.alert.provider.blackduck.descriptor.BlackDuckDescriptor;
 import com.synopsys.integration.alert.util.TestPropertyKey;
@@ -211,7 +210,7 @@ public class EmailChannelChannelDescriptorTestIT extends ChannelDescriptorTest {
 
     @Override
     public TestAction getTestAction() {
-        return new EmailDistributionTestAction(emailChannel, providerDataAccessor, new DefaultEmailHandler(providerDataAccessor));
+        return new EmailDistributionTestAction(emailChannel, new EmailAddressHandler(providerDataAccessor), providerDataAccessor);
     }
 
     @Test
