+++ src/test/java/com/synopsys/integration/alert/channel/util/event/ChannelEventManagerTest.java
@@ -1,4 +1,4 @@
-package com.synopsys.integration.alert.channel.event;
+package com.synopsys.integration.alert.channel.util.event;
 
 import java.util.Date;
 import java.util.List;
@@ -10,6 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.jms.core.JmsTemplate;
 
+import com.synopsys.integration.alert.channel.util.ChannelEventManager;
 import com.synopsys.integration.alert.common.ContentConverter;
 import com.synopsys.integration.alert.common.event.AlertEvent;
 import com.synopsys.integration.alert.common.event.ContentEvent;
