+++ src/test/java/com/synopsys/integration/alert/provider/blackduck/BlackDuckProviderTest.java
@@ -39,7 +39,7 @@ public class BlackDuckProviderTest {
         final TaskManager taskManager = Mockito.mock(TaskManager.class);
         final BlackDuckProperties blackDuckProperties = Mockito.mock(BlackDuckProperties.class);
         Mockito.when(blackDuckProperties.createBlackDuckServerConfigSafely(Mockito.any(IntLogger.class))).thenReturn(Optional.empty());
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, null, taskManager, blackDuckProperties, null);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, taskManager, blackDuckProperties, null);
         provider.initialize();
         Mockito.verify(taskManager, Mockito.times(0)).scheduleCronTask(Mockito.anyString(), Mockito.anyString());
     }
@@ -53,7 +53,7 @@ public class BlackDuckProviderTest {
         final TaskManager taskManager = Mockito.mock(TaskManager.class);
         final BlackDuckProperties blackDuckProperties = Mockito.mock(BlackDuckProperties.class);
         Mockito.when(blackDuckProperties.createBlackDuckServerConfigSafely(Mockito.any(IntLogger.class))).thenReturn(Optional.of(Mockito.mock(BlackDuckServerConfig.class)));
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, null, taskManager, blackDuckProperties, null);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, taskManager, blackDuckProperties, null);
         provider.initialize();
         Mockito.verify(taskManager, Mockito.times(2)).scheduleCronTask(Mockito.anyString(), Mockito.anyString());
     }
@@ -66,7 +66,7 @@ public class BlackDuckProviderTest {
         Mockito.when(projectSyncTask.getTaskName()).thenReturn(BlackDuckProjectSyncTask.TASK_NAME);
         final TaskManager taskManager = Mockito.mock(TaskManager.class);
         final BlackDuckProperties blackDuckProperties = Mockito.mock(BlackDuckProperties.class);
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, null, taskManager, blackDuckProperties, null);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, taskManager, blackDuckProperties, null);
         provider.destroy();
         Mockito.verify(taskManager, Mockito.times(2)).unregisterTask(Mockito.anyString());
     }
@@ -79,7 +79,7 @@ public class BlackDuckProviderTest {
         final BlackDuckProperties blackDuckProperties = Mockito.mock(BlackDuckProperties.class);
 
         final BlackDuckContent blackDuckContent = new BlackDuckContent();
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, blackDuckContent, null, taskManager, blackDuckProperties, null);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, blackDuckContent, taskManager, blackDuckProperties, null);
         final Set<String> expectedNotificationTypes = new LinkedHashSet<>();
         expectedNotificationTypes.add(NotificationType.POLICY_OVERRIDE.name());
         expectedNotificationTypes.add(NotificationType.RULE_VIOLATION.name());
@@ -102,7 +102,7 @@ public class BlackDuckProviderTest {
         final Set<MessageContentCollector> expectedCollectorSet = new HashSet<>(collectorList);
         Mockito.when(topicCollectorFactory.createTopicCollectors()).thenReturn(expectedCollectorSet);
 
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, null, taskManager, blackDuckProperties, topicCollectorFactory);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, null, taskManager, blackDuckProperties, topicCollectorFactory);
         final Set<MessageContentCollector> actualCollectorSet = provider.createTopicCollectors();
         Mockito.verify(topicCollectorFactory).createTopicCollectors();
         assertEquals(expectedCollectorSet, actualCollectorSet);
@@ -116,7 +116,7 @@ public class BlackDuckProviderTest {
         final BlackDuckProperties blackDuckProperties = Mockito.mock(BlackDuckProperties.class);
 
         final BlackDuckContent blackDuckContent = new BlackDuckContent();
-        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, blackDuckContent, null, taskManager, blackDuckProperties, null);
+        final BlackDuckProvider provider = new BlackDuckProvider(accumulatorTask, projectSyncTask, blackDuckContent, taskManager, blackDuckProperties, null);
         final Set<FormatType> expectedNotificationTypes = EnumSet.of(FormatType.DEFAULT, FormatType.DIGEST, FormatType.SUMMARY);
         final Set<FormatType> providerNotificationTypes = provider.getProviderContent().getSupportedContentFormats();
         assertEquals(expectedNotificationTypes, providerNotificationTypes);
